refactor: Register 클래스로 validateDuplicateUser 이동
